refactor: modificacion del nombre del directorio base: de weatherloadercomponent a weather_loader_component
