+++ src/test/java/ar/edu/unq/weather_loader_component/WeatherLoaderComponentApplicationTests.java
@@ -1,4 +1,4 @@
-package ar.edu.unq.weatherloadercomponent;
+package ar.edu.unq.weather_loader_component;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
